@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -28,37 +28,35 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.1
-  flutter_facebook_login: ^3.0.0
 
   # google sign in
 #  firebase_auth: ^0.18.3
-  firebase_auth: ^0.20.0+1
-  google_sign_in: ^4.5.6
-  firebase_core_platform_interface: ^3.0.1
-  cloud_firestore: ^0.16.0
+  firebase_auth: ^1.0.3
+  google_sign_in: ^5.0.1
+  firebase_core_platform_interface: ^4.0.0
+  cloud_firestore: ^1.0.4
 
   # state management
-  provider: ^4.3.2+2
+  provider: ^5.0.0
 
   # icon of google
-  font_awesome_flutter: ^8.10.0
+  font_awesome_flutter: ^9.0.0
 
   # location services
-  geolocator: ^5.3.0
+  geolocator: ^7.0.1
   geocoder: ^0.2.1
 
-  country_list_pick: ^1.0.0+3
-  flutter_country_picker: ^0.1.6
-  intl: ^0.16.1
-  shared_preferences: ^0.5.12+4
-  fluttertoast: ^7.1.6
-  flash: ^1.3.2
-  search_page: ^1.4.0+1
-  uuid: 2.2.2
-  crypto: 2.1.5
-  date_format: ^1.0.9
-  timeago: ">=2.0.27 <3.0.0"
-  url_launcher: ^5.7.5
+  country_list_pick: ^1.0.1+4
+  intl: ^0.17.0
+  shared_preferences: ^2.0.5
+  fluttertoast: ^8.0.3
+  flash: ^1.5.1
+  search_page: ^2.0.0
+  uuid: ^3.0.4
+  crypto: 3.0.0
+  date_format: ^2.0.2
+  timeago: ^3.0.2
+  url_launcher: ^6.0.3
 
 dev_dependencies:
   flutter_test:
