@@ -3,9 +3,9 @@ import 'package:flutter/foundation.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 
 class DisplayInterests with ChangeNotifier{
-  final String id;
-  final String title;
-  final String type;
+  final String? id;
+  final String? title;
+  final String? type;
 
   DisplayInterests({
     this.id,
@@ -15,9 +15,9 @@ class DisplayInterests with ChangeNotifier{
 
   factory DisplayInterests.fromDocument(DocumentSnapshot doc) {
     return DisplayInterests(
-      id: doc.data()['id'],
-      title: doc.data()['title'],
-      type: doc.data()['type'],
+      id: doc.data()!['id'],
+      title: doc.data()!['title'],
+      type: doc.data()!['type'],
     );
   }
 }
