@@ -1,9 +1,9 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 
 class InterestAccount {
-  final String id;
-  final String title;
-  final bool isSelected;
+  final String? id;
+  final String? title;
+  final bool? isSelected;
 
   InterestAccount({
     this.id,
@@ -13,9 +13,9 @@ class InterestAccount {
 
   factory InterestAccount.fromDocument(DocumentSnapshot doc) {
     return InterestAccount(
-      id: doc.data()['id'],
-      title: doc.data()['title'],
-      isSelected: doc.data()['isSelected'],
+      id: doc.data()!['id'],
+      title: doc.data()!['title'],
+      isSelected: doc.data()!['isSelected'],
     );
   }
 }
