@@ -2,9 +2,9 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/foundation.dart';
 
 class SelectRate with ChangeNotifier {
-  final String id;
-  final int rate;
-  final String feedback;
+  final String? id;
+  final int? rate;
+  final String? feedback;
 
   SelectRate({
     this.id,
@@ -15,7 +15,7 @@ class SelectRate with ChangeNotifier {
   factory SelectRate.fromDocument(DocumentSnapshot doc) {
     final d = doc.data();
     return SelectRate(
-      id: d['id'],
+      id: d!['id'],
       rate: d['rate'],
       feedback: d['feedback'],
     );
