@@ -6,9 +6,9 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 
 class SelectMusic with ChangeNotifier {
-  final String id;
-  final String title;
-  bool isSelected;
+  final String? id;
+  final String? title;
+  bool? isSelected;
   final dynamic selects;
 
   SelectMusic({
@@ -20,8 +20,8 @@ class SelectMusic with ChangeNotifier {
 
   factory SelectMusic.fromDocument(DocumentSnapshot doc) {
     return SelectMusic(
-      id: doc.data()['id'],
-      title: doc.data()['title'],
+      id: doc.data()!['id'],
+      title: doc.data()!['title'],
       // isSelected: doc.data()['isSelected'],
       selects: doc['selects'],
     );
@@ -30,21 +30,21 @@ class SelectMusic with ChangeNotifier {
   Future<void> toggleSelectedStatus() async {
     final user = FirebaseAuth.instance.currentUser;
     UserAccount currentUser;
-    currentUser = UserAccount(id: user.uid);
-    String currentUserID = currentUser.id;
+    currentUser = UserAccount(id: user!.uid);
+    String currentUserID = currentUser.id!;
     final musicRef = FirebaseFirestore.instance.collection('music');
     final usersRef = FirebaseFirestore.instance.collection('users');
     // isSelected = !isSelected;
     isSelected = selects[currentUserID] == true;
     try {
-      if (isSelected) {
+      if (isSelected!) {
         await musicRef.doc(id).update({
           'selects.$currentUserID': false,
         });
         selects[currentUserID] = false;
         //....
         await usersRef.doc(currentUser.id).collection('interests').doc(id).delete();
-      } else if (!isSelected) {
+      } else if (!isSelected!) {
         await musicRef.doc(id).update({
           'selects.$currentUserID' : true
         });
