@@ -3,9 +3,9 @@ import 'package:flutter/foundation.dart';
 import 'package:challenge_seekpania/models/selections/select_games.dart';
 
 class GameItem {
-  final String id;
-  final String title;
-  final bool isSelected;
+  final String? id;
+  final String? title;
+  final bool? isSelected;
 
   GameItem({this.id, this.title, this.isSelected});
 }
