@@ -6,7 +6,7 @@ import 'package:challenge_seekpania/models/user_account.dart';
 import 'package:challenge_seekpania/widget/home/activity/activity_details.dart';
 
 class FavoritesItem extends StatefulWidget {
-  UserAccount user;
+  UserAccount? user;
 
   FavoritesItem({this.user});
 
@@ -15,23 +15,23 @@ class FavoritesItem extends StatefulWidget {
 }
 
 class _FavoritesItemState extends State<FavoritesItem> {
-  UserAccount currentUser;
+  late UserAccount currentUser;
   final user = FirebaseAuth.instance.currentUser;
 
   @override
   Widget build(BuildContext context) {
-    currentUser = UserAccount(id: user.uid);
+    currentUser = UserAccount(id: user!.uid);
     return Column(
       children: [
         ListTile(
           leading: CircleAvatar(
-            backgroundImage: NetworkImage(widget.user.photoURL),
+            backgroundImage: NetworkImage(widget.user!.photoURL!),
             maxRadius: 25,
           ),
           title: Padding(
             padding: EdgeInsets.only(left: 10.0),
             child: Text(
-              widget.user.firstName,
+              widget.user!.firstName!,
               style: TextStyle(
                   fontSize: 14.0
                 // color: Color(0xffff3366),
@@ -39,7 +39,7 @@ class _FavoritesItemState extends State<FavoritesItem> {
             ),
           ),
           onTap: () {
-            print(widget.user.id);
+            print(widget.user!.id);
           },
         ),
       ],
