@@ -18,9 +18,9 @@ class FaceToFace extends StatefulWidget {
   // static const routeName = './activity-face-to-face';
   // DUMMY FACE TO FACE
 
-  final String interestID;
-  final String interest;
-  final String type;
+  final String? interestID;
+  final String? interest;
+  final String? type;
 
   FaceToFace({this.interestID, this.interest, this.type});
 
@@ -37,7 +37,7 @@ class _FaceToFaceState extends State<FaceToFace> {
   final _captionForm = GlobalKey<FormState>();
   final _howManyForm = GlobalKey<FormState>();
   // var _editedInvite = SelectInvite(id: null, caption: '', participants: '');
-  String caption, participants;
+  String? caption, participants;
 
   bool isOneSelected = false;
   bool isGroupSelected = false;
@@ -50,17 +50,17 @@ class _FaceToFaceState extends State<FaceToFace> {
   bool isLocation = true;
   bool isInvalidLocation = false;
 
-  String notes;
-  String location;
+  String? notes;
+  String? location;
 
   //--- date and time starts here
 
   bool isNowSelected = false;
   bool isLaterSelected = false;
 
-  String _setTime, _setDate;
-  String _hour, _minute, _time;
-  String dateTime;
+  String? _setTime, _setDate;
+  String? _hour, _minute, _time;
+  String? dateTime;
   DateTime selectedDate = DateTime.now();
   TimeOfDay selectedTime = TimeOfDay(hour: 00, minute: 00);
 
@@ -70,13 +70,13 @@ class _FaceToFaceState extends State<FaceToFace> {
   //--- date and time initialization ends here
 
   Future<Null> _selectDate(BuildContext context) async {
-    final DateTime picked = await showDatePicker(
+    final DateTime picked = (await showDatePicker(
         context: context,
         initialDate: selectedDate,
         initialDatePickerMode: DatePickerMode.day,
         firstDate: DateTime(2021),
         lastDate: DateTime(2101)
-    );
+    ))!;
     if (picked != null)
       setState(() {
         selectedDate = picked;
@@ -86,17 +86,17 @@ class _FaceToFaceState extends State<FaceToFace> {
   }
 
   Future<Null> _selectTime(BuildContext context) async {
-    final TimeOfDay picked = await showTimePicker(
+    final TimeOfDay picked = (await showTimePicker(
       context: context,
       initialTime: selectedTime,
-    );
+    ))!;
     if (picked != null)
       setState(() {
         selectedTime = picked;
         _hour = selectedTime.hour.toString();
         _minute = selectedTime.minute.toString();
-        _time = _hour + ' : ' + _minute;
-        _timeController.text = _time;
+        _time = _hour! + ' : ' + _minute!;
+        _timeController.text = _time!;
         _timeController.text = formatDate(
             DateTime(2019, 08, 1, selectedTime.hour, selectedTime.minute),
             [hh, ':', nn, "" "", am]
@@ -118,8 +118,8 @@ class _FaceToFaceState extends State<FaceToFace> {
 
   //SUBMIT
   Future<void> _submit() async {
-    final isCaptionValid = _captionForm.currentState.validate();
-    final isHowManyValid = _howManyForm.currentState.validate();
+    final isCaptionValid = _captionForm.currentState!.validate();
+    final isHowManyValid = _howManyForm.currentState!.validate();
     if (isOneSelected == false && isGroupSelected == false && participantsController.text.isNotEmpty) {
       setState(() {
         isCompanionSelected = true;
@@ -147,8 +147,8 @@ class _FaceToFaceState extends State<FaceToFace> {
       return;
     }
 
-    _captionForm.currentState.save();
-    _howManyForm.currentState.save();
+    _captionForm.currentState!.save();
+    _howManyForm.currentState!.save();
 
     setState(() {
       isOneSelected = false;
@@ -159,8 +159,8 @@ class _FaceToFaceState extends State<FaceToFace> {
       // isInvalidLocation = false;
     });
 
-    String type;
-    var count = int.parse(participants);
+    String? type;
+    var count = int.parse(participants!);
     if (count == 1) {
       type = 'One Companion';
     } else if (count > 1) {
@@ -270,7 +270,7 @@ class _FaceToFaceState extends State<FaceToFace> {
               decoration: BoxDecoration(
                   borderRadius: BorderRadius.all(Radius.circular(10.0)),
                   border: Border.all(
-                    color: Colors.deepPurple[900],
+                    color: Colors.deepPurple[900]!,
                   )
               ),
               child: Center(
@@ -311,14 +311,14 @@ class _FaceToFaceState extends State<FaceToFace> {
                 fontStyle: FontStyle.italic,
               ),
               enabledBorder: UnderlineInputBorder(
-                borderSide: BorderSide(color: Colors.deepPurple[900]),
+                borderSide: BorderSide(color: Colors.deepPurple[900]!),
               ),
               focusedBorder: UnderlineInputBorder(
-                borderSide: BorderSide(color: Colors.deepPurple[900]),
+                borderSide: BorderSide(color: Colors.deepPurple[900]!),
               ),
             ),
             validator: (value) {
-              if (value.isEmpty) {
+              if (value!.isEmpty) {
                 return 'Please provide a caption';
               }
               return null;
@@ -346,7 +346,7 @@ class _FaceToFaceState extends State<FaceToFace> {
                   height: 60.0,
                   decoration: new BoxDecoration(
                     border: Border.all(
-                      color: isOneSelected ? Colors.deepPurple[900] : Colors.black,
+                      color: isOneSelected ? Colors.deepPurple[900]! : Colors.black,
                     ),
                     color: isOneSelected ? Colors.deepPurple[900] : Colors.white,
                     // shape: BoxShape.circle,
@@ -392,7 +392,7 @@ class _FaceToFaceState extends State<FaceToFace> {
                   height: 60.0,
                   decoration: new BoxDecoration(
                     border: Border.all(
-                      color: isGroupSelected ? Colors.deepPurple[900] : Colors.black,
+                      color: isGroupSelected ? Colors.deepPurple[900]! : Colors.black,
                     ),
                     color: isGroupSelected ? Colors.deepPurple[900] : Colors.white,
                     // shape: BoxShape.circle,
@@ -460,14 +460,14 @@ class _FaceToFaceState extends State<FaceToFace> {
                       fontStyle: FontStyle.italic,
                     ),
                     enabledBorder: UnderlineInputBorder(
-                      borderSide: BorderSide(color: Colors.deepPurple[900]),
+                      borderSide: BorderSide(color: Colors.deepPurple[900]!),
                     ),
                     focusedBorder: UnderlineInputBorder(
-                      borderSide: BorderSide(color: Colors.deepPurple[900]),
+                      borderSide: BorderSide(color: Colors.deepPurple[900]!),
                     ),
                   ),
                   validator: (value) {
-                    if (value.isEmpty) {
+                    if (value!.isEmpty) {
                       return 'Invalid!';
                     }
                     if (int.parse(value) <= 0) {
