@@ -9,7 +9,7 @@ import 'package:challenge_seekpania/widget/interest_box.dart';
 
 class UsersProfile extends StatefulWidget {
 
-  final UserAccount user;
+  final UserAccount? user;
 
   UsersProfile({this.user});
 
@@ -18,7 +18,7 @@ class UsersProfile extends StatefulWidget {
 }
 
 class _UsersProfileState extends State<UsersProfile> {
-  String gender;
+  String? gender;
 
   var _isInit = true;
   var _isLoading = false;
@@ -30,7 +30,7 @@ class _UsersProfileState extends State<UsersProfile> {
       setState(() {
         _isLoading = true;
       });
-      Provider.of<Interest>(context).fetchAllInterests(widget.user.id).then((_) {
+      Provider.of<Interest>(context).fetchAllInterests(widget.user!.id!).then((_) {
         setState(() {
           _isLoading = false;
         });
@@ -55,7 +55,7 @@ class _UsersProfileState extends State<UsersProfile> {
   );
 
   viewInterest() {
-    final interests = Provider.of<Interest>(context, listen: false).userInterests.map((g) => InterestBox(g.title)).toList();
+    final interests = Provider.of<Interest>(context, listen: false).userInterests.map((g) => InterestBox(g.title!)).toList();
     // final liveEvents = Provider.of<LiveEvents>(context, listen: false).eventItems.map((g) => InterestBox(g.title)).toList();
     // final interests = games;
 
@@ -79,10 +79,10 @@ class _UsersProfileState extends State<UsersProfile> {
     // print('THE NEW PHOTO URL');
     // print(result);
 
-    if (widget.user.gender == 'Non-Binary') {
-      gender = widget.user.genderCustom;
+    if (widget.user!.gender == 'Non-Binary') {
+      gender = widget.user!.genderCustom!;
     } else {
-      gender = widget.user.gender;
+      gender = widget.user!.gender!;
     }
 
     return Scaffold(
@@ -101,7 +101,7 @@ class _UsersProfileState extends State<UsersProfile> {
                         bottomLeft: Radius.circular(20),
                       ),
                       child: Image.network(
-                        widget.user.photoURL,
+                        widget.user!.photoURL!,
                         // 's400-c/$userPhoto',
                         // 'https://lh3.googleusercontent.com/a-/AOh14GjZwVeyNas_d37ucE1zyFcth-1b33CYoXU8lEUj=s400-c',
                         width: 390,
@@ -120,7 +120,7 @@ class _UsersProfileState extends State<UsersProfile> {
                         ),
                         children: <TextSpan>[
                           TextSpan(
-                            text: widget.user.firstName,
+                            text: widget.user!.firstName,
                             style: TextStyle(
                               fontSize: 30,
                               fontWeight: FontWeight.bold,
@@ -136,7 +136,7 @@ class _UsersProfileState extends State<UsersProfile> {
                             ),
                           ),
                           TextSpan(
-                            text: widget.user.age.toString(),
+                            text: widget.user!.age.toString(),
                             style: TextStyle(
                                 fontSize: 18,
                             ),
@@ -160,7 +160,7 @@ class _UsersProfileState extends State<UsersProfile> {
                             ),
                           ),
                           TextSpan(
-                            text: widget.user.status,
+                            text: widget.user!.status,
                             style: TextStyle(
                               fontSize: 18,
                             ),
@@ -172,7 +172,7 @@ class _UsersProfileState extends State<UsersProfile> {
                             ),
                           ),
                           TextSpan(
-                            text: widget.user.city,
+                            text: widget.user!.city,
                             style: TextStyle(
                               fontSize: 18,
                             ),
