@@ -17,9 +17,9 @@ import 'package:challenge_seekpania/widget/virtual/virtual_set_date_time.dart';
 class VirtualGathering extends StatefulWidget {
   // static const routeName = './activity-face-to-face';
 
-  final String interestID;
-  final String interest;
-  final String type;
+  final String? interestID;
+  final String? interest;
+  final String? type;
 
   VirtualGathering({this.interestID, this.interest, this.type});
 
@@ -36,7 +36,7 @@ class _VirtualGatheringState extends State<VirtualGathering> {
   final _captionForm = GlobalKey<FormState>();
   final _howManyForm = GlobalKey<FormState>();
   // var _editedInvite = SelectInvite(id: null, caption: '', participants: '');
-  String caption, participants;
+  String? caption, participants;
 
   bool isOneSelected = false;
   bool isGroupSelected = false;
@@ -49,17 +49,17 @@ class _VirtualGatheringState extends State<VirtualGathering> {
   bool isLocation = true;
   bool isInvalidLocation = false;
 
-  String notes;
-  String location;
+  String? notes;
+  String? location;
 
   //--- date and time starts here
 
   bool isNowSelected = false;
   bool isLaterSelected = false;
 
-  String _setTime, _setDate;
-  String _hour, _minute, _time;
-  String dateTime;
+  String? _setTime, _setDate;
+  String? _hour, _minute, _time;
+  String? dateTime;
   DateTime selectedDate = DateTime.now();
   TimeOfDay selectedTime = TimeOfDay(hour: 00, minute: 00);
 
@@ -69,13 +69,13 @@ class _VirtualGatheringState extends State<VirtualGathering> {
   //--- date and time initialization ends here
 
   Future<Null> _selectDate(BuildContext context) async {
-    final DateTime picked = await showDatePicker(
+    final DateTime picked = (await showDatePicker(
         context: context,
         initialDate: selectedDate,
         initialDatePickerMode: DatePickerMode.day,
         firstDate: DateTime(2021),
         lastDate: DateTime(2101)
-    );
+    ))!;
     if (picked != null)
       setState(() {
         selectedDate = picked;
@@ -85,17 +85,17 @@ class _VirtualGatheringState extends State<VirtualGathering> {
   }
 
   Future<Null> _selectTime(BuildContext context) async {
-    final TimeOfDay picked = await showTimePicker(
+    final TimeOfDay picked = (await showTimePicker(
       context: context,
       initialTime: selectedTime,
-    );
+    ))!;
     if (picked != null)
       setState(() {
         selectedTime = picked;
         _hour = selectedTime.hour.toString();
         _minute = selectedTime.minute.toString();
-        _time = _hour + ' : ' + _minute;
-        _timeController.text = _time;
+        _time = _hour! + ' : ' + _minute!;
+        _timeController.text = _time!;
         _timeController.text = formatDate(
             DateTime(2019, 08, 1, selectedTime.hour, selectedTime.minute),
             [hh, ':', nn, "" "", am]
@@ -116,8 +116,8 @@ class _VirtualGatheringState extends State<VirtualGathering> {
   }
 
   Future<void> _submit() async {
-    final isCaptionValid = _captionForm.currentState.validate();
-    final isHowManyValid = _howManyForm.currentState.validate();
+    final isCaptionValid = _captionForm.currentState!.validate();
+    final isHowManyValid = _howManyForm.currentState!.validate();
     if (isOneSelected == false && isGroupSelected == false && participantsController.text.isNotEmpty) {
       setState(() {
         isCompanionSelected = true;
@@ -159,8 +159,8 @@ class _VirtualGatheringState extends State<VirtualGathering> {
     //   return;
     // }
 
-    _captionForm.currentState.save();
-    _howManyForm.currentState.save();
+    _captionForm.currentState!.save();
+    _howManyForm.currentState!.save();
 
     setState(() {
       isOneSelected = false;
@@ -171,8 +171,8 @@ class _VirtualGatheringState extends State<VirtualGathering> {
       // isInvalidLocation = false;
     });
 
-    String type;
-    var count = int.parse(participants);
+    String? type;
+    var count = int.parse(participants!);
     if (count == 1) {
       type = 'One Companion';
     } else if (count > 1) {
@@ -211,8 +211,8 @@ class _VirtualGatheringState extends State<VirtualGathering> {
             // ActivitySearch(searchID: widget.interestID, searchType: widget.type, caption: caption, participants: participants)
             // ActivitySearch(searchID: widget.interestID, searchType: widget.type, activity: _editedActivity)
             VirtualSetDateTime(
-              searchID: widget.interestID, interestName: widget.interest, searchType: widget.type,
-              caption: caption, companionType: type, participants: count,
+              searchID: widget.interestID!, interestName: widget.interest!, searchType: widget.type!,
+              caption: caption!, companionType: type!, participants: count,
             )
         )
     );
@@ -283,7 +283,7 @@ class _VirtualGatheringState extends State<VirtualGathering> {
               decoration: BoxDecoration(
                   borderRadius: BorderRadius.all(Radius.circular(10.0)),
                   border: Border.all(
-                    color: Colors.deepPurple[900],
+                    color: Colors.deepPurple[900]!,
                   )
               ),
               child: Center(
@@ -324,14 +324,15 @@ class _VirtualGatheringState extends State<VirtualGathering> {
                 fontStyle: FontStyle.italic,
               ),
               enabledBorder: UnderlineInputBorder(
-                borderSide: BorderSide(color: Colors.deepPurple[900]),
+                borderSide: BorderSide(color: Colors.deepPurple[900]!,
+                ),
               ),
               focusedBorder: UnderlineInputBorder(
-                borderSide: BorderSide(color: Colors.deepPurple[900]),
+                borderSide: BorderSide(color: Colors.deepPurple[900]!),
               ),
             ),
             validator: (value) {
-              if (value.isEmpty) {
+              if (value!.isEmpty) {
                 return 'Please provide a caption';
               }
               return null;
@@ -340,9 +341,9 @@ class _VirtualGatheringState extends State<VirtualGathering> {
             onSaved: (value) {
               caption = value;
             },
-          ),
         ),
       ),
+     )
     );
   }
 
@@ -359,7 +360,7 @@ class _VirtualGatheringState extends State<VirtualGathering> {
                   height: 60.0,
                   decoration: new BoxDecoration(
                     border: Border.all(
-                      color: isOneSelected ? Colors.deepPurple[900] : Colors.black,
+                      color: isOneSelected ? Colors.deepPurple[900]! : Colors.black,
                     ),
                     color: isOneSelected ? Colors.deepPurple[900] : Colors.white,
                     // shape: BoxShape.circle,
@@ -405,7 +406,7 @@ class _VirtualGatheringState extends State<VirtualGathering> {
                   height: 60.0,
                   decoration: new BoxDecoration(
                     border: Border.all(
-                      color: isGroupSelected ? Colors.deepPurple[900] : Colors.black,
+                      color: isGroupSelected ? Colors.deepPurple[900]! : Colors.black,
                     ),
                     color: isGroupSelected ? Colors.deepPurple[900] : Colors.white,
                     // shape: BoxShape.circle,
@@ -473,14 +474,14 @@ class _VirtualGatheringState extends State<VirtualGathering> {
                       fontStyle: FontStyle.italic,
                     ),
                     enabledBorder: UnderlineInputBorder(
-                      borderSide: BorderSide(color: Colors.deepPurple[900]),
+                      borderSide: BorderSide(color: Colors.deepPurple[900]!),
                     ),
                     focusedBorder: UnderlineInputBorder(
-                      borderSide: BorderSide(color: Colors.deepPurple[900]),
+                      borderSide: BorderSide(color: Colors.deepPurple[900]!),
                     ),
                   ),
                   validator: (value) {
-                    if (value.isEmpty) {
+                    if (value!.isEmpty) {
                       return 'Invalid!';
                     }
                     if (int.parse(value) <= 0) {
