@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 
 AppBar header(context,
-    {bool isAppTitle = false, String titleText, removeBackButton = false}) {
+    {bool isAppTitle = false, String? titleText, removeBackButton = false}) {
   return AppBar(
     automaticallyImplyLeading: removeBackButton ? false : true,
     title: Text(
-      isAppTitle ? "Seekpania" : titleText,
+      (isAppTitle ? "Seekpania" : titleText)!,
       style: TextStyle(
         color: Colors.white,
         fontFamily: isAppTitle ? "Signatra" : "",
