@@ -21,16 +21,16 @@ class _EditGameScreenState extends State<EditGameScreen> {
   var _isLoading = false;
 
   Future<void> _saveForm() async {
-    final isValid = _form.currentState.validate();
+    final isValid = _form.currentState!.validate();
     if (!isValid) {
       return;
     }
-    _form.currentState.save();
+    _form.currentState!.save();
     setState(() {
       _isLoading = true;
     });
     if (_editedGame.id != null) {
-      await Provider.of<Games>(context, listen: false).updateGame(_editedGame.id, _editedGame);
+      await Provider.of<Games>(context, listen: false).updateGame(_editedGame.id!, _editedGame);
     } else {
       try {
         await Provider.of<Games>(context, listen: false).addGame(_editedGame);
@@ -48,11 +48,11 @@ class _EditGameScreenState extends State<EditGameScreen> {
   void didChangeDependencies() {
     if (_isInit) {
       // gameId IS NULL FIX THIS!
-      final gameId = ModalRoute.of(context).settings.arguments as String;
+      final gameId = ModalRoute.of(context)!.settings.arguments as String;
       if (gameId != null) {
         _editedGame = Provider.of<Games>(context, listen: false).findById(gameId);
         _initValues = {
-          'title': _editedGame.title,
+          'title': _editedGame.title.toString(),
         };
       }
     }
@@ -88,7 +88,7 @@ class _EditGameScreenState extends State<EditGameScreen> {
                 ),
                 textInputAction: TextInputAction.next,
                 validator: (value) {
-                  if (value.isEmpty) {
+                  if (value!.isEmpty) {
                     return 'Please provide the name of the game.';
                   }
                   return null;
